ECOSORTBIN/requirements.txt
absl-py==2.0.0
astunparse==1.6.3
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
flatbuffers==23.5.26
gast==0.5.4
google-api-core==2.15.0
google-auth==2.25.1
google-auth-oauthlib==1.1.0
google-cloud-core==2.4.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.6.0
googleapis-common-protos==1.62.0
greenlet==3.0.1
grpcio==1.60.0
gunicorn==20.1.0
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.15.0
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
mysql-connector-python==8.2.0
mysqlclient==2.2.0
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
protobuf==4.21.12
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.23
tensorboard==2.15.1
tensorflow==2.15.0
tensorflow-estimator==2.15.0
termcolor==2.4.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0